CaseTecJus/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
openai>=1.51.0
numpy==1.26.4
requests==2.32.3
streamlit==1.39.0
python-dotenv==1.0.1

